refactor: Update function names and data encoding in encode_fq.py

+++ src/types.rs
@@ -1,7 +1,7 @@
 use ndarray::{Array2, Array3};
 use std::collections::HashMap;
 
-pub type Element = i64;
+pub type Element = i32;
 pub type Matrix = Array2<Element>;
 pub type Tensor = Array3<Element>;
 pub type Kmer2IdTable = HashMap<Vec<u8>, Element>;
